+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Snow Walker trait. It allows the player to walk on powder snow when the trait is on either armor or a curio.
+  - Found on leather, wool, and fine silk, but is restricted to boots.
+
 ## [2.8.3] - 2022-01-07
 ### Fixed
 - Wild flax and fluffy plants not generating or being very rare [#457]